Integrando rota `/carrinho` + Fluxo entre telas home-produto-carrinho
