+++ ProjetoRUP/musicshop-backend/src/comunicacao/TelaCarrinhoPresenter.ts
@@ -1,0 +1,32 @@
+import { Request, Response } from "express";
+import { injectable } from "tsyringe";
+
+import Fachada from "../negocio/Fachada/Fachada";
+
+@injectable()
+class TelaCarrinhoPresenter {
+  private fachada;
+
+  constructor(fachada: Fachada) {
+    this.fachada = fachada;
+  }
+
+  public pegarCarrinho(req: Request, res: Response) {
+    const clienteId = req.body.clienteId;
+    const carrinho = this.fachada.pegarCarrinho(clienteId);
+    if (carrinho) {
+      res.json(carrinho);
+    } else {
+      throw new Error(
+        "Não foi possível encontrar o carrinho do cliente " + clienteId
+      );
+    }
+  }
+
+  public criarPedido(req: Request, res: Response) {
+    this.fachada.criarPedido(req.body.cliente);
+    res.status(201).send();
+  }
+}
+
+export default TelaCarrinhoPresenter;
