--- ProjetoRUP/musicshop-backend/src/comunicacao/TelaPedidoPresenter.ts
@@ -1,20 +1,0 @@
-import { Request, Response } from "express";
-import { injectable } from "tsyringe";
-
-import Fachada from "../negocio/Fachada/Fachada";
-
-@injectable()
-class TelaPedidoPresenter {
-  private fachada;
-
-  constructor(fachada: Fachada) {
-    this.fachada = fachada;
-  }
-
-  public criarPedido(req: Request, res: Response) {
-    this.fachada.criarPedido(req.body.cliente);
-    res.status(201).send();
-  }
-}
-
-export default TelaPedidoPresenter;
